refactor: remove metadata mod
