+++ src/common/meta/src/key/flow_task/flow_task_name.rs
@@ -18,7 +18,8 @@ use serde::{Deserialize, Serialize};
 use snafu::OptionExt;
 
 use crate::error::{self, Result};
-use crate::key::scope::{CatalogScoped, FlowTaskScoped, MetaKey};
+use crate::key::flow_task::FlowTaskScoped;
+use crate::key::scope::{CatalogScoped, MetaKey};
 use crate::key::txn_helper::TxnOpGetResponseSet;
 use crate::key::{
     txn_helper, DeserializedValueWithBytes, FlowTaskId, TableMetaValue, NAME_PATTERN,
